# packages/yucca-admin-api/openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/oidc/login": {
      "get": {
        "operationId": "oidcAuthorize",
        "parameters": [
          {
            "name": "code_challenge",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "yucca-admin",
    "description": "yucca admin API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
...
}